fix test, move objects to legit destinations
