+++ src/app/pieces.py
@@ -1,8 +1,11 @@
+import re
 from math import inf as infinity
 from string import ascii_lowercase
-from typing import Union, Tuple, List, Iterator
+from typing import Union, Tuple, List, Iterator, Optional
 
-from domain.pieces import Piece, Side, Position, location_regex
+from domain.pieces import Piece, Side, Position
+
+location_regex = re.compile(r'^(?P<file>[a-zA-Z]+)(?P<rank>[0-9]+)$')
 
 
 class King(Piece):
@@ -236,3 +239,38 @@ class StandardPosition(Position):
             return self.rank
         else:
             raise IndexError("tuple index out of range")
+
+
+class Move:  # TODO: Next to abstract? fix constructor
+    """
+    Two Position aggregator with optional pawn promotion information
+    """
+
+    def __init__(self, a: Position, b: Position, promotion: Optional[Piece] = None):
+        self.__a = a
+        self.__b = b
+        self.__promotion = promotion
+
+    @property
+    def a(self):
+        return self.__a
+
+    @property
+    def b(self):
+        return self.__b
+
+    @property
+    def promotion(self):
+        return self.__promotion
+
+    def __repr__(self):
+        if self.__promotion:
+            return 'Move: %s to %s with promotion to %s' % (self.__a, self.__b, self.__promotion.name)
+        else:
+            return 'Move: %s to %s' % (self.__a, self.__b)
+
+    def __str__(self):
+        if self.__promotion:
+            return '%s%s%s' % (self.__a, self.__b, self.__promotion.char)
+        else:
+            return '%s%s' % (self.__a, self.__b)
